requirements.txt
# Core Dependencies
elasticsearch>=8.0.0
elasticsearch-dsl>=8.0.0
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=0.19.0
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
sqlalchemy>=2.0.0
alembic>=1.8.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
python-dateutil>=2.8.0
pytz>=2022.1

# Threat Intelligence
pymisp>=2.4.0
taxii2-client>=2.3.0
stix2>=3.0.0
feedparser>=6.0.0

# Communication & Alerts
aiosmtplib>=2.0.0
slack-sdk>=3.21.0
discord.py>=2.1.0
python-telegram-bot>=20.0

# Monitoring & Logging
loguru>=0.6.0
prometheus-client>=0.15.0
structlog>=22.1.0

# Testing & Development
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Security
cryptography>=3.4.0
bcrypt>=4.0.0
python-jose>=3.3.0

# Background Tasks
celery>=5.2.0
redis>=4.3.0

# Web Interface (Optional)
streamlit>=1.15.0
plotly>=5.11.0
dash>=2.7.0